+++ R/bb_helpers.R
@@ -21,7 +21,7 @@
 #'   package = c("billboarder", "ggiraph", "officer",
 #'               "shinyWidgets", "visNetwork", "rAmCharts", 
 #'               "D3partitionR"),
-#'   stars = c(36, 194, 72, 61, 183, 25, 18)
+#'   stars = c(67, 252, 160, 144, 224, 32, 25)
 #' )
 #' 
 #' # By default, first column is mapped on the x-axis
@@ -50,7 +50,7 @@
 #'   package = c("billboarder", "ggiraph", "officer",
 #'               "shinyWidgets", "visNetwork", "rAmCharts", 
 #'               "D3partitionR"),
-#'   stars = c(36, 194, 72, 61, 183, 25, 18)
+#'   stars = c(67, 252, 160, 144, 224, 32, 25)
 #' )
 #' 
 #' billboarder() %>%
@@ -326,6 +326,16 @@ bb_categories <- function(bb, categories) {
 #'     data = iris, 
 #'     mapping = bbaes(Sepal.Length, Sepal.Width, group = Species)
 #'   )
+#'   
+#' # Size variable
+#' billboarder() %>% 
+#'   bb_scatterplot(
+#'     data = iris, 
+#'     mapping = bbaes(Sepal.Length, Sepal.Width,
+#'                     group = Species, size = Petal.Width),
+#'     range = c(0.5, 120)
+#'   ) %>% 
+#'   bb_x_axis(tick = list(fit = FALSE))
 #'
 bb_scatterplot <- function(bb, data, mapping = NULL, ...) {
   
@@ -760,11 +770,13 @@ bb_histogram <- function(bb, x, breaks = "Sturges", ...) {
 #' @param bb A \code{billboard} \code{htmlwidget} object.
 #' @param data A \code{data.frame} or a \code{vector}.
 #' @param mapping Mapping of variables on the chart, see \code{\link{bbaes}}.
-#' @param type Type of chart : line, spline, step, area, area-spline, area-step.
+#' @param type Type of chart : line, spline, step, area, area-spline, area-step, 
+#'  area-line-range, area-spline-range.
 #' @param show_point Whether to show each point in line.
 #' @param ... Not used.
 #' 
-#' @note This function can be used with \code{\link{billboarderProxy}} in shiny application.
+#' @note Types area-line-range and area-spline-range don't work in RStudio viewer, open chart in a browser. 
+#'  This function can be used with \code{\link{billboarderProxy}} in shiny application.
 #'
 #' @return A \code{billboard} \code{htmlwidget} object.
 #' @export
@@ -847,11 +859,41 @@ bb_histogram <- function(bb, x, breaks = "Sturges", ...) {
 #' billboarder() %>% 
 #'   bb_linechart(data = lynx.df, x = "year")
 #'   
+#'   
+#' ### Area range charts
+#' 
+#' # Generate data
+#' dat <- data.frame(
+#'   date = seq.Date(Sys.Date(), length.out = 20, by = "day"),
+#'   y1 = round(rnorm(20, 100, 15)),
+#'   y2 = round(rnorm(20, 100, 15))
+#' )
+#' dat$ymin1 <- dat$y1 - 5
+#' dat$ymax1 <- dat$y1 + 5
+#' 
+#' dat$ymin2 <- dat$y2 - sample(3:15, 20, TRUE)
+#' dat$ymax2 <- dat$y2 + sample(3:15, 20, TRUE)
+#' 
+#' 
+#' # Make chart : use ymin & ymax aes for range
+#' billboarder(data = dat) %>% 
+#'   bb_linechart(
+#'     mapping = bbaes(x = date, y = y1, ymin = ymin1, ymax = ymax1),
+#'     type = "area-line-range"
+#'   ) %>% 
+#'   bb_linechart(
+#'     mapping = bbaes(x = date, y = y2, ymin = ymin2, ymax = ymax2), 
+#'     type = "area-spline-range"
+#'   ) %>% 
+#'   bb_y_axis(min = 50)
+#'   
 bb_linechart <- function(bb, data, mapping = NULL, type = "line", show_point = FALSE, ...) {
   
   type <- match.arg(
     arg = type, 
-    choices = c("line", "spline", "step", "area", "area-spline", "area-step", "bar"),
+    choices = c("line", "spline", "step", "area", 
+                "area-spline", "area-step", "bar",
+                "area-line-range", "area-spline-range"),
     several.ok = TRUE
   )
   
@@ -874,6 +916,7 @@ bb_linechart <- function(bb, data, mapping = NULL, type = "line", show_point = F
   } else {
     if (!is.null(mapping)) {
       data <- bbmapping(data = data, mapping = mapping)
+      # if (!is.null(bb$data$json$y))
     } 
     if (inherits(x = data[[1]], what = c("Date", "POSIXct"))) {
       if (inherits(x = data[[1]], what = c("POSIXct"))) {
@@ -888,20 +931,20 @@ bb_linechart <- function(bb, data, mapping = NULL, type = "line", show_point = F
       } else {
         data[[1]] <- as.character(data[[1]])
       }
-      
-      data_opt <- list(
-        x = names(data)[1],
-        json = as.list(data),
-        type = type
-      )
       if (!"billboarder_Proxy" %in% class(bb)) {
         bb <- bb_x_axis(bb, type = "timeseries")
       }
-    } else {
+    }
+    if (length(data) > 1) {
       data_opt <- list(
         x = names(data)[1],
         json = as.list(data),
-        type = type
+        types = setNames(as.list(rep_len(type, length(data) - 1)), nm = names(data)[-1])
+      )
+    } else {
+      data_opt <- list(
+        json = as.list(data),
+        types = setNames(list(type), nm = names(data)[1])
       )
     }
   }